requirements.txt
Pillow
pandas
gspread
oauth2client
requests